refacto import from base empreinte
